refactoring: minor moving/restructuring
