+++ src/scripts/Module/constants.ts
@@ -1,1 +1,4 @@
 export const MODULE_KEY = 'focus-sheet'
+
+export const SETTING_AUTO_OPEN = 'auto-open'
+export const SETTING_AUTO_OPEN_POPOUT = 'auto-open-popout'
