+++ src/scripts/Popout/constants.ts
@@ -1,3 +1,1 @@
 export const MODULE_KEY = 'popout'
-export const SETTING_AUTO_OPEN = 'auto-open'
-export const SETTING_AUTO_OPEN_POPOUT = 'auto-open-popout'
